+++ package.json
@@ -4,15 +4,18 @@
   "version": "0.1.1",
   "private": true,
   "dependencies": {
+    "ajv": "^6.12.0",
     "bootstrap": "^4.4.1",
+    "core-js": "^3.6.4",
     "gh-pages": "^1.2.0",
-    "react": "^16.4.1",
-    "react-dev-utils": "^10.2.0",
-    "react-dom": "^16.4.2",
+    "jquery": "^3.4.1",
+    "react": "^16.13.1",
+    "react-dev-utils": "^10.2.1",
+    "react-dom": "^16.13.1",
     "react-router-dom": "^4.3.1",
-    "react-scripts": "1.1.4",
-    "react-scroll": "^1.7.10",
-    "reactstrap": "^6.3.0"
+    "react-scripts": "^3.4.1",
+    "react-scroll": "^1.7.16",
+    "reactstrap": "^6.5.0"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -21,5 +24,17 @@
     "predeploy": "npm run build",
     "deploy": "gh-pages -d build",
     "eject": "react-scripts eject"
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
   }
 }
