+++ src/App.js
@@ -1,14 +1,11 @@
 import React, { Component } from 'react';
-import './App.css';
+//import './App.css';
 
 import { Button } from 'reactstrap';
 
 import Navigationbar from './components/Navigationbar';
-import Landing from './components/Landing';
-import About from './components/About';
-import Projects from './components/Projects';
+import Landing from './pages/Landing';
 
-import Section from './components/Section';
 
 class App extends Component {
   render() {
@@ -18,10 +15,7 @@ class App extends Component {
         <header className="App-header">
           <Navigationbar />
         </header>
-        <Section />
         <Landing/>
-        <About/>
-        <Projects/>
       </div>
 
     );
