--- src/components/Section.js
@@ -1,20 +1,0 @@
-import React from 'react';
-import { Container, Row, Col, Button } from 'reactstrap';
-
-import Landing from './Landing';
-
-// ES6 Imports
-import * as Scroll from 'react-scroll';
-import { Link, Element , Events, animateScroll as scroll, scrollSpy, scroller } from 'react-scroll'
- 
-
-export default class Section extends React.Component {
-
-    render() {
-        return (
-            <div>
-
-            </div>
-        );
-    }
-}
