+++ src/pages/styles/landing-default.css
@@ -1,0 +1,38 @@
+#align-center {
+  text-align: center;
+}
+#social-but-links {
+  padding-top: 3vmax;
+  flex-wrap: wrap;
+}
+#welcome-info-body {
+  margin-top: 7vmax;
+  max-width: 720px;
+}
+#welcome-info-text {
+}
+#title {
+  padding-top: 72px;
+  font-size: 6.93vmax;
+  font-family: Impact, Haettenschweiler, 'Arial Narrow Bold', sans-serif;
+  color: rgb(102, 252, 241);
+}
+#subtitle {
+  font-size: 3.69vmax;
+  font-family: Impact, Haettenschweiler, 'Arial', sans-serif;
+  color: rgb(212, 212, 212);
+}
+#jumbo {
+  background-color: rgb(31, 40, 51);
+}
+#landing-view {
+  background-color: rgb(31, 40, 51);
+  color: white;
+}
+
+/* Gradient transparent - color - transparent */
+#line-breaker {
+  border: 0;
+  height: 1px;
+  background-image: linear-gradient(to right, rgba(255, 255, 255, 0), rgba(255, 255, 255, 0.75), rgba(255, 255, 255, 0));
+}
